requirements.txt
RPi.GPIO==0.6.1
crcmod==1.7
spidev==3.0
